+++ src/vonage/errors.py
@@ -52,9 +52,13 @@ class ProactiveConnectError(ClientError):
     """An error relating to the Proactive Connect API."""
 
 
-class NumberInsightV2Error(ClientError):
-    """An error relating to the Number Insight v2 API."""
-
-
 class InvalidPhoneNumberError(ClientError):
     """An invalid phone number was supplied."""
+
+
+class UsersError(ClientError):
+    """An error relating to the Users API."""
+
+
+class NumberInsightV2Error(ClientError):
+    """An error relating to the Number Insight v2 API."""