+++ src/vonage/redact.py
@@ -2,7 +2,11 @@ from .errors import RedactError
 
 from deprecated import deprecated
 
-@deprecated(version='3.0.0', reason='This is a dev preview product and as such is not supported in this SDK.')
+
+@deprecated(
+    version='3.0.0',
+    reason='This is a dev preview product and as such is not supported in this SDK.',
+)
 class Redact:
     auth_type = 'header'
 
@@ -16,7 +20,9 @@ class Redact:
         params = {"id": id, "product": product}
         if type is not None:
             params["type"] = type
-        return self._client.post(self._client.api_host(), "/v1/redact/transaction", params, auth_type=Redact.auth_type)
+        return self._client.post(
+            self._client.api_host(), "/v1/redact/transaction", params, auth_type=Redact.auth_type
+        )
 
     def _check_allowed_product_name(self, product):
         if product not in self.allowed_product_names: