+++ tests/test_getters_setters.py
@@ -2,6 +2,7 @@ def test_getters(client, dummy_data):
     assert client.host() == dummy_data.host
     assert client.api_host() == dummy_data.api_host
 
+
 def test_setters(client, dummy_data):
     try:
         client.host('host.vonage.com')